feat(worker): added --ondemand worker support
With current node it is only possible to implement --ondemand
worker profiling.
