+++ lib/Runner.js
@@ -5,10 +5,9 @@
  */
 
 const childProcess = require('child_process');
-const EventEmitter = require('events');
 const readline = require('readline');
 
-const createSocket = require('./InspectorWebSocket');
+const { WebSocketTransport, WorkerTransport, Connection } = require('./Connection');
 
 class ReportWrapper {
   constructor(id, callback) {
@@ -75,35 +74,27 @@ async function runTool({nodeCommandLine, toolFactory, ondemand, callback}) {
   await new Promise(resolve => nodeProcess.once('exit', resolve));
 }
 
-function setupEnvironment() {
-  const nativeConsoleDebug = console.debug;
-  function sendCommand(command, filename) {
-    let resolveCallback;
-    const promise = new Promise(resolve => resolveCallback = resolve);
-    const id = setTimeout(resolve, 30000);
-    function resolve() {
-      resolveCallback();
-      clearTimeout(id);
+async function setupOnDemand(connection, wrapper) {
+  function setupEnvironment() {
+    const nativeConsoleDebug = console.debug;
+    function sendCommand(command) {
+      let resolveCallback;
+      const promise = new Promise(resolve => resolveCallback = resolve);
+      const id = setTimeout(resolve, 30000);
+      function resolve() {
+        resolveCallback();
+        clearTimeout(id);
+      }
+      nativeConsoleDebug(command, resolve);
+      return promise;
     }
-    nativeConsoleDebug(command, resolve, filename);
-    return promise;
-  }
-  this.startTheTool = sendCommand.bind(null, 'thetool:start');
-  this.stopTheTool = sendCommand.bind(null, 'thetool:stop');
-}
-
-async function runToolForProcess({webSocket, toolFactory, ondemand, callback}) {
-  const ws = await createSocket(webSocket);
-  const connection = new Connection(ws);
-  const wrapper = new ToolWrapper(toolFactory(connection), callback);
-  if (ondemand) {
-    await connection.send('Runtime.evaluate', {
-      expression: `(${setupEnvironment.toString()})()`
-    });
-  } else {
-    await wrapper.start();
+    this.startTheTool = sendCommand.bind(null, 'thetool:start');
+    this.stopTheTool = sendCommand.bind(null, 'thetool:stop');
   }
 
+  await connection.send('Runtime.evaluate', {
+    expression: `(${setupEnvironment.toString()})()`
+  });
   connection.on('Runtime.consoleAPICalled', async message => {
     if (message.type !== 'debug')
       return;
@@ -121,17 +112,47 @@ async function runToolForProcess({webSocket, toolFactory, ondemand, callback}) {
       objectId: message.args[1].objectId
     });
   });
+}
+
+async function runToolForProcess({webSocket, toolFactory, ondemand, callback}) {
+  const connection = new Connection(await WebSocketTransport.create(webSocket));
+  const tool = toolFactory(connection);
+  const wrapper = new ToolWrapper(tool, callback);
+  if (ondemand) {
+    await setupOnDemand(connection, wrapper);
+    await connection.send('NodeWorker.enable', {
+      waitForDebuggerOnStart: true
+    });
+    connection.on('NodeWorker.attachedToWorker', ({sessionId}) => runToolForWorker({
+      connection,
+      sessionId,
+      toolFactory,
+      callback,
+      wrapper: tool.onlyMain() ? wrapper : null
+    }));
+  } else {
+    await wrapper.start();
+  }
 
   await connection.send('Runtime.enable');
   let finishedCallback;
   const finishedPromise = new Promise(resolve => finishedCallback = resolve);
   connection.on('Runtime.executionContextDestroyed', message => message.executionContextId === 1 && finishedCallback());
+
   await connection.send('Runtime.runIfWaitingForDebugger');
   await finishedPromise;
 
   if (!ondemand)
     await wrapper.stop();
   await connection.close();
+}
+
+async function runToolForWorker({connection: mainConnection, sessionId, toolFactory, wrapper, callback}) {
+  const connection = new Connection(new WorkerTransport(mainConnection, sessionId));
+  wrapper = wrapper || new ToolWrapper(toolFactory(connection), callback);
+  await setupOnDemand(connection, wrapper);
+  await connection.send('Runtime.enable');
+  await connection.send('Runtime.runIfWaitingForDebugger', {});
 }
 
 const blacklist = new Set([
@@ -165,62 +186,6 @@ function waitForWSEndpoint(nodeProcess, callback) {
     }
     callback(match[1]);
   }
-}
-
-class Connection extends EventEmitter {
-  constructor(ws) {
-    super();
-    this._lastId = 0;
-    this._ws = ws;
-    this._callbacks = new Map();
-    this._ws.on('message', this._onMessage.bind(this));
-  }
-
-  async send(method, params = {}) {
-    const id = this._rawSend({method, params});
-    return new Promise((resolve, reject) => {
-      this._callbacks.set(id, {resolve, error: new Error(), reject, method});
-    });
-  }
-
-  _rawSend(message) {
-    const id = ++this._lastId;
-    message = JSON.stringify(Object.assign({}, message, {id}));
-    this._ws.send(message);
-    return id;
-  }
-
-  async _onMessage(message) {
-    const object = JSON.parse(message);
-    if (object.id) {
-      const callback = this._callbacks.get(object.id);
-      if (callback) {
-        this._callbacks.delete(object.id);
-        if (object.error)
-          callback.reject(createProtocolError(callback.error, callback.method, object));
-        else
-          callback.resolve(object.result);
-      }
-    } else {
-      this.emit(object.method, object.params);
-    }
-  }
-
-  async close() {
-    this._ws.close();
-  }
-}
-
-function createProtocolError(error, method, object) {
-  let message = `Protocol error (${method}): ${object.error.message}`;
-  if ('data' in object.error)
-    message += ` ${object.error.data}`;
-  return rewriteError(error, message);
-}
-
-function rewriteError(error, message) {
-  error.message = message;
-  return error;
 }
 
 module.exports = runTool;
