+++ lib/Tool.js
@@ -22,6 +22,10 @@ class MemorySamplingProfiler {
     reporter.reportChunk(JSON.stringify(data.profile));
     reporter.reportFinish();
   }
+
+  onlyMain() {
+    return false;
+  }
 }
 
 class MemoryAllocationProfiler {
@@ -43,6 +47,10 @@ class MemoryAllocationProfiler {
     this._listener = null;
     reporter.reportFinish();
   }
+
+  onlyMain() {
+    return false;
+  }
 }
 
 class CoverageProfiler {
@@ -62,6 +70,10 @@ class CoverageProfiler {
     reporter.reportChunk(JSON.stringify(data));
     reporter.reportFinish();
   }
+
+  onlyMain() {
+    return false;
+  }
 }
 
 class CPUProfiler {
@@ -79,6 +91,10 @@ class CPUProfiler {
     const data = await this._connection.send('Profiler.stop');
     reporter.reportChunk(JSON.stringify(data.profile));
     reporter.reportFinish();
+  }
+
+  onlyMain() {
+    return false;
   }
 }
 
@@ -99,6 +115,10 @@ class TypeProfiler {
     reporter.reportChunk(JSON.stringify(data));
     reporter.reportFinish();
   }
+
+  onlyMain() {
+    return false;
+  }
 }
 
 class Tracing {
@@ -118,13 +138,13 @@ class Tracing {
       console.log('Available categories: ', categories);
       return;
     }
-    reporter.reportStart('json', 'You can use Chrome DevTools Performance tab to load and analyze it.');
     await this._connection.send('NodeTracing.start', {
       traceConfig: {
         recordMode: this._recordMode,
         includedCategories: this._includedCategories
       }
     });
+    reporter.reportStart('json', 'You can use Chrome DevTools Performance tab to load and analyze it.');
     let first = true;
     this._listener = message => reportChunk(message.value);
     this._connection.on('NodeTracing.dataCollected', this._listener);
@@ -146,6 +166,10 @@ class Tracing {
       reporter.reportChunk(']');
       reporter.reportFinish();
     }
+  }
+
+  onlyMain() {
+    return true;
   }
 }
